start our own http server instead of using ifrit's
this ensures the rep will always exit during evacuation even if it ran into the
following conditions:

1. garden `Destroy' hangs during evacuation
2. the rep receives a `/stop' request

the rep http request will hang trying to acquire the opLock in the
StoreNode. prior to this change, this would prevent the process to exit as
`ifrit.httpServer' will wait for all http requests to finish gracefully before
it shuts down.

[#154486141]

Signed-off-by: John Shahid <jvshahid@gmail.com>