wip: second proposal for retryable tasks
in this proposal the BBS keep track of the retry/failure counts. only allowing
3 retries. Rep doesn't use the FailTask at all and FailTask usage is limited to
the Auctioneer. BBS treat FailTask as a retrytable failure while a CompleteTask
is not.

Note: we could use FailTask in the Rep for example to retry tasks if their
container disappear. I think that's an interesting benefit of this apprach.

[#155348160]

Signed-off-by: John Shahid <jvshahid@gmail.com>