Fix flaky after each for main test
The fake `/containers` endpoint was erroneously returning
"healthcheck-container" for the task containers when the block channel
was finally closed in the AfterEach. This change prevents the
`/containers` endpoint from returning the wrong handle for the task
containers, instead opting to just end the request handling when the
block channel finishes.