Add support for ECR images
When user provides a docker URL that matches ECR repo the username and
password are used as AWS access key id and secret key to obtain
authorization token that is passed to Garden to pull the image.

aws.Session is constructed instead of ECRHElper because ECRHelper only
allows AWS access key id and secret key to be passed as environment
variables which is not acceptable in multithreaded host process. Besides
that the code follows ECRHelper implementation to obtain authorization
token:
https://github.com/awslabs/amazon-ecr-credential-helper/blob/master/ecr-login/ecr.go

[#168209772](https://www.pivotaltracker.com/story/show/168209772)