Move AfterEach logic out of goroutine
Most of the code in there includes calls to Eventually and Wait, so we don't need
the goroutine/done channel logic.

The only possible issue is fakeGarden.Close, which includes code that acquires/releases
a mutex. If we find that tests are hanging on that line, we'll need to wrap that single
line in a goroutine managed by a done channel